lect 47- seperate component to store files
