+++ src/App.jsx
@@ -1,34 +1,8 @@
-import reactImg from './assets/react-core-concepts.png';
 import componentImage from './assets/components.png';
 import stateImage from './assets/state-mgmt.png';
 import { CORE_CONCEPTS } from './data.js'
-const reactDescription = ['Fundamnetal', 'Crucial', 'Core', 'Significant'];
-
-function generateIndex(max) {
-  return Math.floor(Math.random() * (max + 1));
-}
-
-function CoreConcept(props) {
-  return (
-    <li>
-    <img src={props.img} alt ="componentImage" />
-    <h3>{props.tile}</h3>
-    <p>{props.descrtion}</p>
-  </li>
-  )
-}
-
-function Header() {
-  return (<header>
-    <img src={reactImg} alt="Stylized atom" />
-    <h1>React Essentials</h1>
-    <p>
-      {/* Output Dynamic data */}
-      {reactDescription[generateIndex(3)]} React concepts you will need for almost any app you are
-      going to build!
-    </p>
-  </header>);
-}
+import Header from './Components/Header.jsx';
+import CoreConcept from './Components/CoreConcept.jsx';
 
 function App() {
   return (
